refactor(fan): extract fan-mode constants to const.py, move MideaFanOnlyFanEntity to climate.py
